// interfaces/product.ts
export interface Product {
  ProductId: number;
  Description: string;
  Cost: number;
  Price: number;
  Sku: string;
  Name: string;
  LargePhotoSrc: any;
  ThumbnailPhotoSrc: any;
  ThumbnailPhotoFileName: string;
  LargePhotoFileName: string;
  ProductSubcategoryId: number;
  Weight: number;
  Size: string;
  WeightUnitMeasureCode: string;
  SizeUnitMeasureCode: string;
  LineCount: number;
  ProductCategoryId: number;
  ProductModelId: number;
  Color: string;
}